passes primitive types by value, ODEs [cxx]
COMMENTS:
It is really not necessary to pass primitive types by reference
since there won't be any impact on the performance of the code.

The original intention was to make it evident that those constants
won't be changed by any of the methods and to have the compiler
check that for me. I don't think it's necessary to continue
doing that at this point, especially because the methods require
few lines of code so it's easy to see that those parameters do
not change.

The method declarations looked unduly long for my taste ...
I am not entirely satisfied with the aesthetics yet but it's
slighly better.