fixes iteration count, nonlinear solvers
COMMENTS:
In most implementations the first iteration is carried out before
the while-loop, so the counter should not start at zero as it was
done.

The report needs only to check if the iteration count is not equal
to the maximum number of iterations. Note that if the method fails
to obtain the solution the iteration counter is guaranteed to be
equal to the maximum number of iterations. Also note that in most
implementations the logical AND is short-circuited so that if the
counter is equal to the maximum number of iterations the while
loop will stop without bothering to check if f(x) satisfies the
convergence criterion.

Runtime checks were done to verify that the report is generated
accordingly.