implements the quick norm method [clang]
COMMENTS:
The method skips the computation of the squared root for speed.

Rationale:
This method is meant for the computation of the norm of the error
vector. When the norm of the error vector is smaller than the
tolerance the linear solver stops because the solution has been
found.

Note that one only needs to use the squared of the intended
tolerance; that is, use a tolerance of 1.0e-12 instead of 1.0e-6.