prepares the path for solving PDEs [clang]
COMMENTS:
Differentiates the right-hand side RHS of the PDE from the source
term.

It so happens that the RHS and the source are equal for the
steady problem (ODE) being used to test the linear solvers.

In a future commit it will become evident that the right-hand side
of the PDE has a contribution from the transient terms in addition
to the source term.

To differentiate the right-hand side from the source term the
array of vectors is incremented by one to accomodate the source
term. The right-hand side (odevector) is still passed as an argument
to the iterative linear solver. The change is that the solver is
tasked with the initialization of the right-hand side (odevector).
This approach will also be used for solving the transient problem.

This commit is preparing the way for the changes to be applied.