adds support for solving PDEs [clang]
COMMENTS:
This is another step towards adding support for solving PDEs.

Defines the parameter `alpha' = (`dx' * `dx') / `dt', where `dx'
is the spatial step along the x-axis and `dt' is the temporal step.

To solve the steady problem we simply use `alpha' = 0 and to solve
the transient problem we can choose a suitable value for it O(1).

We are currently using zero `alpha' because we are still solving
the steady problem.

This parameter is now part of the iterative solver parameters,
see the changes done to the header file.

Note that the linear solvers now initialize the right-hand side
(odevec) with both the source term and the transient terms -- see
the for-loop where we update the finite-difference vector. As you
may recall, on the previous commit we had the solvers copy the
source term into the odevec to initialize it. Now we are adding
the transient term to complete the computation of odevec. It does
not hurt to do that for the steady problem because the transient
terms are proportional to `alpha' so we are adding a bunch of
zeros. (In terms of performance it does not hurt to do so because
the time complexity is determined by the main loop of the solvers.)

Also note that when solving the transient problem the diagonal
elements of the tridiagonal system of equations acquires a
transient contribution which is also proportional to `alpha'.
So instead of `2' the diagonal becomes `alpha + 2', it is easy
to see that we recover the steady problem when `alpha' is zero.
This is reflected in the definition of the new constant `c' of
double type just before the main loop of the linear solvers.

Numeric results are unaffected as it should be.

Valgrind reports no memory leak issues.

Code shall be added to solve the transient problem in a future
commit.