adds support for silencing the solver [clang]
COMMENTS:
Adds code that can silence the linear solvers. It would be a waste
of CPU cycles to print on the standard stream output the number of
iterations every time.

Now the users can suppress the solver output if they want.

Note that when solving the transient problem the linear solvers
will be called on every time step so it is a good idea to silence
them for the reasons mentioned above.