passes solver params by reference [clang]
COMMENTS:
Passes the solver parameter struct by reference instead of by
copying.

Since the solver parameter is declared with the `const' qualifier a
reference to a constant (solver parameter) type is defined as the
argument of the solver -- see the prototypes.

The reason for this commit is just a matter of style, not a
performance related modification for the cost of passing a `small'
struct such as this one is small. We are just promising that the
(solver) parameter struct will not be altered by the linear solver
itself, as it should be.