removes heap allocations from solvers [clang]
COMMENTS:
This is another step towards adding support for solving PDEs.

Removes allocations on the heap from the iterative linear solvers
for performance reasons.

Note that the guess and the error vectors are no longer local
variables but passed as arguments, for these are now elements of
the vector array (pdesol).

An advantage of doing this is that we can use the guess vector
from the previous (time) step to solve the current step when
dealing with the transient problem. Of course there are other ways
to achieve the same feat but this one is good enough without
additional effort.

Numeric results are unaffected by the change.

Valgrind reports no memory issues.