creates the state vector [clang]
COMMENTS:
This is another step towards adding support for solving PDEs.

If the state vector holds a failed state value we know that we
should stop integrating the PDE with respect time and try again
with different solver parameters.

It is worth mentioning that the code should simply skip post
processing tasks if the solver returns a failed state. No need to
halt execution of the application in a hurry so that we get a
chance to free the alloted resources gracefully (to avert memory
leaks).

This is the rationale for adding the state vector.

Numeric results are unaffected as it should be.

Valgrind reports no memory issues.