fixes the time computation [clang]
COMMENTS:
Note that after the call to the solver (Jacobi or Gauss-Seidel) we
obtain the time t = (i + 1) * dt, where i = 0 the first time the
loop executes so t = dt, as it should be. The last value of i is
(`steps' - 1) just before the loop breaks so the final time must
be t = steps * dt. This is why we now set the final time with that
value.