exports the fields obtained by PDE solvers [clang]
COMMENTS:
Exports both the numeric and analytic (temperature) fields f(t, x)
if the solver succeeds.

Note that if an input-output IO error occurs the code simply skips
the export instead of halting execution right away to get a chance
to free the allocated resources (on the heap for the vectors) for
the sake of terminating execution gracefully without incurring on
memory leaks.