does post-processing when solver is successful
COMMENTS:
Creates a switch for post-processing so that this task is only
done when the solver has successfully integrated the PDE up to
the sought time step.

To achieve this the (existing) post-processing code has been moved
to the body of the pertaining if-conditional. No other changes have
been made.