exports the field g(t,x) with respect time [clang]
COMMENTS:
Exports the field every 256 time steps, this applies to the
Gauss-Seidel based solver. The final time step has been increased
to obtain a temperature profile g(t, x) near the steady state for
subsequent visualization (in a future commit).

The field data g(t, x) is written to the same data file in a row
wise fashion. The time is the first data element that is written
and then the field g(t, x) for convenience. This means that each
record in the data file consists of `size' + 1 elements, where
`size' is the number of elements in the vector that stores the
field g(t, x) and the extra one is for the `time'.

Note that this makes it easy to get both the time `t' and the field
g(t, x) when processing the contents of the data file.

Also note that the time is defined as t = i * dt because the export
is done before calling the iterative solver (which gets the field
g(t, x) at the next time t = (i + 1) * dt). This means that the
initial field g(t=0, x) is written in the first line of the data
file. This is done because the initial field is going to be used
to prepare a figure that shows g(t, x) for different times in a
future commit.