improves the transient field plot [python]
COMMENTS:
Plots both the analytic and numeric profiles g(t, x) on the same
figure for ease of comparison. It is evident from the plots that
the numeric and analytic solutions are in excellent agreement.

Note that for the analytic solution to describe the initial field
g(t=0, x) it is necessary to use a high order Fourier series; this
is why the order `N' is set to 512.