vectorizes transient 1D Poisson solver [clang]
COMMENTS:
Shows that the restrict keyword does prevent the gcc compiler from
producing different versions of the for-loops to account for
possible array aliasings. Something that FORTRAN programmers do not
need to worry about in general :)

The solver integrates the 1D Poisson equation until steady-state.

There is excellent agreement between the exact and numeric solutions.