validates 1d transient solution Poisson [clang]
COMMENTS:
Validates the transient solution by computing the difference of the
numeric approximation of the temperature field against the exact.

The console application logs the difference between the exact and
numeric solutions while integrating the finite difference equations.

The number of integration steps have been set so that the change
from the transient to the steady state can be deduced from the logs.

The solver tolerance have been reduced further. Recall that we do
not compute the squared root of the norm of the error vector for
speed. This is why it might seem exceendingly small.

Applies the initial and boundary conditions BCs:

	g(t = 0, x) = 1                 (initial condition)
	g(t, x = -1) = g(t, x = 1) = 0  (boundary conditions)

Note that the steady-state solution is independent of the initial
state of the system, that is why there was excellent agreement
between the numeric and steady-state solutions in the previous
commit.