adds auto-vectorized Poisson2D solver [clang]
COMMENTS:
GCC auto-vectorizes some but not all of the important loops.
Nevertheless, there are other strategies that can be applied to
speed up the code like parallelization (via OpenMP or even POSIX
threads).

What truly matters is that the solver describes the transient and
steady-state behavior of the system. The numeric solution agrees
with the analytic solution.

The solver might be modified to export the transient results to a
plain text file for visualization with Python.

Valgrind reports no memory issues.