uses bitmask for auto-vectorization [clang]
COMMENTS:
test.c:
Rewrite loops in a way that GCC can auto-vectorize via bitmasking.

This technique can easily be applied to solve the 2D and 3D Poisson
equation and to other problems with complex geometries.

Numeric results are in agreement with the analytic counterpart.

Valgrind reports no illegal memory issues.

make-inc:
Ignores incompatible pointer type conversion when optimizing code.

Normally I would not ignore this type of warning, especially during
development. However, the ``incompatible'' pointer type conversion
is being done on purpose to write auto-vectorizable loops by GCC
by means of bitmasking of double-precision floating point numbers.

These loops had simple nested conditionals that prevented auto
vectorization at least by GCC. The workaround is to down the
computation into steps by bitmasking and applying whatever operation
that should follow.

Note that the size of the union and the data are the same (64-bits)
so it is ``safe'' to do this given that the compiler now optimizes
the code further.